+++ src/components/Attributes/AttributeViewer/index.tsx
@@ -98,7 +98,7 @@ export default function AttributeViewer({
 
     const getAttributesTableData = useCallback(
         (attribute: AttributeResponseModel | MetadataItemModel) => ({
-            id: attribute.uuid || "",
+            id: attribute.uuid || attribute.name,
             columns: [
                 attribute.label || attribute.name || "",
                 getEnumLabel(contentTypeEnum, attribute.contentType),
