+++ src/components/_pages/custom-attributes/detail/index.tsx
@@ -137,8 +137,8 @@ export default function CustomAttributeDetail() {
                           id: "resources",
                           columns: [
                               "Resources",
-                              customAttribute.resources?.map((r) => (
-                                  <Badge key={r} style={{ margin: "1px" }} color="secondary">
+                              customAttribute.resources?.map((r, i) => (
+                                  <Badge key={i} style={{ margin: "1px" }} color="secondary">
                                       {getEnumLabel(resourceEnum, r)}
                                   </Badge>
                               )) ?? "",
