+++ src/utils/common-utils.ts
@@ -20,3 +20,7 @@ export const removeNullValues = (obj: any): any => {
 
     return obj;
 };
+
+export const capitalize = (str: string): string => {
+    return str.charAt(0).toUpperCase() + str.slice(1);
+};
